10. Надо подкачаться. Исправил замечания наставника
