+++ js/main.js
@@ -1,5 +1,5 @@
 import { changePrice, syncCheckTime, adFormHandler } from './form.js';
-import { getData } from './map.js';
+import { getData } from './backend.js';
 
 adFormHandler(document.querySelector('.ad-form'));
 changePrice();
