+++ js/map.js
@@ -101,55 +101,22 @@ formAdress.onfocus = function () {
 // Добавляем главную метку на карту
 mainMarker.addTo(map);
 
-const ALERT_SHOW_TIME = 5000;
-
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 100;
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = 0;
-  alertContainer.style.top = 0;
-  alertContainer.style.right = 0;
-  alertContainer.style.padding = '30px 3px';
-  alertContainer.style.fontSize = '30px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = '#FF5454';
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
+// функция добавления меток на карту
+const markerMaker = (mapPoints) => {
+  mapPoints.forEach(function ({ author, offer, location }) {
+    const regularMarker = L.marker(
+      {
+        lat: location.lat,
+        lng: location.lng,
+      },
+      {
+        icon: regularMapMarker,
+      },
+    );
+    regularMarker
+      .addTo(map)
+      .bindPopup(generateOffer({ author, offer }), { keepInView: true });
+  });
 };
 
-// Получаем данные с сервера и добавляем метки на карту
-function getData() {
-  const markerMaker = (mapPoints) => {
-    mapPoints.forEach(function ({ author, offer, location }) {
-      const regularMarker = L.marker(
-        {
-          lat: location.lat,
-          lng: location.lng,
-        },
-        {
-          icon: regularMapMarker,
-        },
-      );
-      regularMarker
-        .addTo(map)
-        .bindPopup(generateOffer({ author, offer }), { keepInView: true });
-    });
-  };
-
-  fetch('https://22.javascript.pages.academy/keksobooking/data')
-    .then((response) => response.json())
-    .then((dataFromServer) => {
-      markerMaker(dataFromServer);
-    })
-    .catch(() => {
-      showAlert('Ошибка загрузки меток. Попробуйте ещё раз');
-    });
-}
-
-export { mapBlock, mapUnBlock, getData };
+export { mapBlock, mapUnBlock, markerMaker };
